dont cast gate lora to fp32

+++ vllm/model_executor/models/glm4_moe.py
@@ -226,10 +226,10 @@ class Glm4MoE(nn.Module):
         hidden_states = hidden_states.view(-1, hidden_dim)
 
         # router_logits: (num_tokens, n_experts)
-        router_logits = self.gate(hidden_states.to(dtype=torch.float32))
+        router_logits = self.gate(hidden_states)
 
         fused_moe_out = self.experts(
-            hidden_states=hidden_states, router_logits=router_logits
+            hidden_states=hidden_states, router_logits=router_logits.to(dtype=torch.float32)
         )
 
         if self.shared_experts is not None:
